vehicle create, get, get specific vehicle, update, delete successfully

+++ src/modules/user/user.service.ts
@@ -1,23 +1,9 @@
 import { pool } from "../../config/db";
 import bcrypt from "bcryptjs";
 
-const createUser = async (payload: Record<string, unknown>) => {
-  const { name, role, email, password } = payload;
-
-  const hashedPass = await bcrypt.hash(password as string, 10);
-
-  const result = await pool.query(
-    `
-      INSERT INTO users (name, role, email, password) VALUES ($1, $2, $3, $4) RETURNING *`,
-    [name, role, email, hashedPass]
-  );
-
-  return result;
-};
-
 const getUsers = async () => {
   const result = await pool.query(`
-      SELECT * FROM users
+      SELECT id, name, email, phone, role FROM users
       `);
 
   return result;
@@ -26,7 +12,7 @@ const getUsers = async () => {
 const getUser = async (id: string) => {
   const result = await pool.query(
     `
-      SELECT * FROM users WHERE id = $1
+      SELECT id, name, email, phone, role, created_at, updated_at FROM users WHERE id = $1
   `,
     [id]
   );
@@ -37,7 +23,7 @@ const getUser = async (id: string) => {
 const updateUser = async (name: string, email: string, id: string) => {
   const result = await pool.query(
     `
-      UPDATE users SET name = $1, email = $2 WHERE id = $3 RETURNING *
+      UPDATE users SET name = $1, email = $2 WHERE id = $3 RETURNING id, name, email, phone, role, created_at, updated_at
       `,
     [name, email, id]
   );
@@ -48,7 +34,7 @@ const updateUser = async (name: string, email: string, id: string) => {
 const deleteUser = async (id: string) => {
   const result = await pool.query(
     `
-      DELETE FROM users WHERE id = $1 RETURNING *
+      DELETE FROM users WHERE id = $1 RETURNING id, name, email, phone, role, created_at, updated_at
       `,
     [id]
   );
@@ -57,7 +43,6 @@ const deleteUser = async (id: string) => {
 };
 
 export const userServices = {
-  createUser,
   getUsers,
   getUser,
   updateUser,
